Обновил загрузку с срм с файлами.

+++ src/attachments/attachments.service.ts
@@ -5,23 +5,79 @@ import { InjectModel } from '@nestjs/sequelize';
 import { nanoid } from 'nanoid';
 import * as path from 'path';
 import * as fs from 'fs';
-import { CreateMessageDto } from 'src/messages/dto/create-message.dto';
+import { AttachmentsProvider } from './attachments.provider';
 
 @Injectable()
 export class AttachmentsService {
-    constructor(@InjectModel(Attachment) private attachmentRepository: typeof Attachment) {}
+    constructor(
+        @InjectModel(Attachment) private attachmentRepository: typeof Attachment,
+        private readonly attachmentsProvider: AttachmentsProvider,
+    ) {}
 
-    async createAttachment(dto: CreateAttachmentDto, filePath) {
+    async sortAttachments(message_id, attachments, messenger_type, message_from) {
+        let attachmentsDataArray;
+
+        switch (messenger_type) {
+            case 'telegram':
+            /* attachmentsDataArray =
+                    await this.attachmentsProvider.telegramAttachmentsParser(attachments);
+                break; */
+            case 'wa':
+                break;
+            case 'vk':
+                attachmentsDataArray = await this.attachmentsProvider.vkAttachmentsParser(
+                    attachments,
+                    message_from,
+                );
+                break;
+            case 'crm':
+                return;
+            default:
+                console.error(`Неизвестный тип мессенджера: ${messenger_type}`);
+                return [];
+        }
+
+        // Если массив пустой, возвращаем пустой результат
+        if (!attachmentsDataArray || attachmentsDataArray.length === 0) {
+            console.warn('Массив вложений пуст.');
+            return [];
+        }
+
+        // Проходимся по каждому элементу массива и вызываем this.CreateAttachments
+        const createdAttachments = [];
+        for (const attachmentData of attachmentsDataArray) {
+            try {
+                // Вызываем вашу функцию CreateAttachments для каждого элемента
+                const createdAttachment = await this.createAttachment({
+                    message_id,
+                    ...attachmentData, // Передаем данные вложения
+                });
+                if (createdAttachment) {
+                    createdAttachments.push(createdAttachment);
+                }
+            } catch (error) {
+                console.error(`Ошибка при создании вложения: ${error.message}`);
+            }
+        }
+
+        // Возвращаем массив созданных вложений
+        return createdAttachments;
+    }
+
+    async createAttachment(dto: CreateAttachmentDto, filePath?) {
         try {
             dto.attachment_id = nanoid();
 
             const attachment = await this.attachmentRepository.create(dto);
-            console.log('attachment успешно создался', attachment);
+            return attachment;
         } catch (error) {
             console.log(error);
-            if (fs.existsSync(filePath)) {
-                fs.unlinkSync(filePath);
+            if (filePath) {
+                if (fs.existsSync(filePath)) {
+                    fs.unlinkSync(filePath);
+                }
             }
+
             throw new HttpException(
                 'Произошла ошибка при сохранении файла-картинки в БД attachments',
                 HttpStatus.INTERNAL_SERVER_ERROR,
